+++ commands/album_commands.py
@@ -139,14 +139,18 @@ class Album:
         card = await self.bot.db.users.get_card_from_album(user.id, card_id)
         valid = False
         if card:
-            valid = (not idolized and card['unidolized_count'] > 0)
-            valid = valid or (idolized and card['idolized_count'] > 0)
+            unidolized_count = card['unidolized_count']
+            idolized_count = card['idolized_count']
+
+            if idolized:
+                valid = (idolized_count > 0)
+            else:
+                valid = (idolized_count > 0 or unidolized_count > 0)
 
         if valid:
-            card_info = await self.bot.db.cards.get_card(card_id)
-            img_url = 'http:' + card_info['card_image']
+            img_url = 'http:' + card['card_image']
             if idolized and card:
-                img_url = 'http:' + card_info['card_idolized_image']
+                img_url = 'http:' + card['card_idolized_image']
 
             fname = basename(urlsplit(img_url).path)
             image_path = idol_img_path.joinpath(fname)
@@ -184,8 +188,7 @@ class Album:
             await self.bot.db.users.add_to_user_album(
                     user.id, [card], idolized=True)
 
-            card_info = await self.bot.db.cards.get_card(card_id)
-            img_url = 'http:' + card_info['card_idolized_image']
+            img_url = 'http:' + card['card_idolized_image']
             fname = basename(urlsplit(img_url).path)
             image_path = idol_img_path.joinpath(fname)
             image = await get_one_img(