+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 - The bot is no longer dependent on School Idol Tomodachi being online
 - !stats and !botstats are new commands (work in progress)
+- You can now view idolized cards in unidolized mode even if you don't have any unidolized copies
 - Fixed a bug where scouts with small search spaces would have a bias for certain cards
 - Fixed sorting by release date of cards
 - Stopped the bot from crashing whenever it fails to pull new card information from SIT